requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.23
pyodbc==5.0.1
python-dotenv==1.0.0